AI_Trading_Website/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.1
numpy==1.26.0
websockets==12.0
openai==0.27.8